Imputing values from log-normal distribution for non-detects

Imputing values from log-normal distribution for non-detects

Usage

impute_censored_with_truncated_fit(data, var, var_cens, distribution = "lnorm")

Arguments

data

dataset of contamination variables

var

character name of the variable containing uncensored data (LOQ values)

var_cens

character name of the variable containing censoring information (T/F)

distribution

character model distribution (default: "lnorm")

Value

vector containing imputed values when data was <LOQ

Examples

impute_censored_with_truncated_fit(data = contam_PFAS_ng_gdw, 
                                   var = "PFNA", var_cens = "PFNA_cen")
#>  [1]  0.80153858  0.75167952  0.24868073  0.41077351  0.24192520  0.16835957
#>  [7]  0.18166090  0.19755872  0.56772974  0.35727140  0.48358491  0.63645384
#> [13]  0.43822938  0.58102827  0.34957467  0.34466321  4.58926475 14.16061297
#> [19]  0.51848130  0.57278802  0.71713831  0.44982346  0.81311606  0.68152412
#> [25]  0.91845414  1.16815333  0.02766662  0.35687875  0.40157384  0.30215792
#> [31]  0.87310279  1.25607166  2.84289658  3.01398939  0.17850955  0.02804539
#> [37]  0.25314977  0.02896430  0.01945613  0.24670991  0.21966194  0.21021898
#> [43]  0.24059560  0.28343835  0.57194570  0.34326063  2.02488780  0.80729368
#> [49]  1.54952243  1.59280831  0.96259352  1.14369899  1.00770791  2.01570792
#> [55]  2.46023857  6.71667092  1.20473459  0.81709917  0.69038692  1.09286506